requirements.txt
python-telegram-bot[webhooks]==20.7
python-docx==1.1.0
aiohttp
